requirements.txt
matplotlib==2.1.0
numpy==1.12.1
ordereddict==1.1
pandas==0.23.3
Pillow==4.2.1
sphinx-argparse==0.2.5
torch==0.4.0
torchvision==0.2.1